+++ src/main/java/com/picimako/terra/psi/js/JSArgumentUtil.java
@@ -1,4 +1,4 @@
-//Copyright 2023 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+//Copyright 2024 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
 package com.picimako.terra.psi.js;
 
@@ -89,13 +89,9 @@ public final class JSArgumentUtil {
      */
     @Nullable
     public static Object getNumericValueOf(JSProperty property) {
-        if (property.getValue() instanceof JSLiteralExpression) {
-            var literal = (JSLiteralExpression) property.getValue();
-            if (literal.isNumericLiteral()) {
-                return literal.getValue();
-            }
-        }
-        return null;
+        return property.getValue() instanceof JSLiteralExpression literal && literal.isNumericLiteral()
+            ? literal.getValue()
+            : null;
     }
 
     private JSArgumentUtil() {