+++ src/main/java/com/picimako/terra/wdio/screenshot/TerraFunctionalTestingScreenshotNameResolver.java
@@ -1,10 +1,10 @@
-//Copyright 2023 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+//Copyright 2024 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
 package com.picimako.terra.wdio.screenshot;
 
-import com.intellij.json.psi.JsonPsiUtil;
 import com.intellij.lang.javascript.psi.JSExpression;
 import com.intellij.lang.javascript.psi.JSLiteralExpression;
+import com.intellij.openapi.util.text.StringUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -60,7 +60,7 @@ public class TerraFunctionalTestingScreenshotNameResolver extends AbstractScreen
     @NotNull
     public String resolveName(JSLiteralExpression element) {
         if (element != null) {
-            String resolved = normalize(parseTestId(JsonPsiUtil.stripQuotes(element.getText())));
+            String resolved = normalize(parseTestId(StringUtil.unquoteString(element.getText())));
             if (!resolved.isEmpty()) {
                 return resolved + ".png";
             }