+++ src/main/java/com/picimako/terra/wdio/screenshot/gutter/TerraScreenshotValidationLineMarkerProvider.java
@@ -1,4 +1,4 @@
-//Copyright 2023 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+//Copyright 2024 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
 package com.picimako.terra.wdio.screenshot.gutter;
 
@@ -41,8 +41,7 @@ public class TerraScreenshotValidationLineMarkerProvider extends RelatedItemLine
 
     @Override
     protected void collectNavigationMarkers(@NotNull PsiElement element, @NotNull Collection<? super RelatedItemLineMarkerInfo<?>> result) {
-        if (isUsingTerra(element.getProject()) && element instanceof JSCallExpression && isScreenshotValidationCall((JSCallExpression) element)) {
-            var terraCallExpr = (JSCallExpression) element;
+        if (isUsingTerra(element.getProject()) && element instanceof JSCallExpression terraCallExpr && isScreenshotValidationCall(terraCallExpr)) {
             var nameArgument = getFirstArgumentAsStringLiteral(terraCallExpr.getArgumentList());
             if (isUsingTerraToolkit(element.getProject()) && nameArgument == null) {
                 var methodExpression = terraCallExpr.getMethodExpression();