+++ src/main/java/com/picimako/terra/wdio/toolwindow/action/CompareLatestWithReferenceScreenshotsAction.java
@@ -1,4 +1,4 @@
-//Copyright 2023 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+//Copyright 2024 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
 package com.picimako.terra.wdio.toolwindow.action;
 
@@ -41,7 +41,7 @@ public class CompareLatestWithReferenceScreenshotsAction extends AbstractTerraWd
         if (project!= null && tree != null && isScreenshot(tree.getLastSelectedPathComponent())) {
             var latests = asScreenshot(tree.getLastSelectedPathComponent()).getLatests();
             if (!latests.isEmpty()) {
-                var fileToOpen = latests.get(0);
+                var fileToOpen = latests.getFirst();
                 var fileEditorManager = FileEditorManager.getInstance(project);
                 fileEditorManager.openFile(fileToOpen, true, fileEditorManager.isFileOpen(fileToOpen));
                 fileEditorManager.setSelectedEditor(fileToOpen, ReferenceToLatestScreenshotsPreview.EDITOR_TYPE_ID);